pyproject.toml
[project]
name = "devops-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.43",
]
